assignments/class_ass_context_handoff_tools/pyproject.toml
[project]
name = "class-ass-context-handoff-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
